ChangePointMean

Documentation for ChangePointMean.

ChangePointMean.mcpall

ChangePointMean.mcplast

ChangePointMean.mcplot

ChangePointMean.mcpoint

ChangePointMean.mcppv

ChangePointMean.mcptime

ChangePointMean.ssq

ChangePointMean.mcpall — Method

mcpall(ts::AbstractVector; pcut = 0.05, shuffle = 500, minlen = 3)

return vector of index of all change points in ts via recursive partitioning pcut is p-value threshold of existence of a change point, default is 0.05 shuffle is number of randomizations to use minlen is minimum allowed length for a segment (prevent short segments)

ChangePointMean.mcplast — Method

mcplast(ts::AbstractVector; pcut=0.05, shuffle=500, minlen=3)

return index of rightmost change point in ts if any, 0 otherwise pcut is p-value threshold of existence of a change point, default is 0.05 shuffle is number of randomizations to use minlen is minimum allowed length for a segment (prevent short segments)

ChangePointMean.mcplot — Method

mcplot(ts::AbstractVector; chgpts=Int[], palette=:seaborn_bright)

returns plot of ts with chgpts as change points (default is none) palette is the default color scheme to use

mcplot(ts) is a plain plot of the time series, mcplot(ts, chgpts=mcpall(ts)) is plot of ts with all change points

ChangePointMean.mcpoint — Method

mcpoint(ts::AbstractVector; pcut=0.05, shuffle=500, minlen=3)

return index of change point in mean in ts if it exists, 0 otherwise pcut is p-value threshold of existence of a change point, default is 0.05 shuffle is number of randomizations to use minlen is minimum allowed length for a segment (prevent short segments)

ChangePointMean.mcppv — Method

mcppv(ts::AbstractVector; shuffle::Int = 500)

return p-value of a change point in mean in ts shuffle is number of randomization used to generate p-value uses global random number generator

ChangePointMean.mcptime — Method

mcptime(ts::AbstractVector; minlen::Int = 1)

return index of start of new segment via minimum ssq in ts minlen is minimum length of a segment you want 0 is returned if minlen is not least twice length of ts

ChangePointMean.ssq — Method

ssq(v) returns sum of squares